:hankey: Fix error from broken status endpoint
For now the error has been suppressed, and a FIXME left.
